Crate raft_proto

Modules

eraftpb
Generated file from eraftpb.proto

prelude

util

Traits

ConfChangeI
Abstracts over ConfChangeV2 and (legacy) ConfChange to allow treating them in a unified manner.

Functions

conf_state_eq

new_conf_change_single
Creates a ConfChangeSingle.

parse_conf_change
Parses a Space-delimited sequence of operations into a slice of ConfChangeSingle. The supported operations are:

stringify_conf_change
The inverse to parse_conf_change.